\name{agstrat}
\alias{agstrat}
\title{Data from a stratified random sample of size 300 from...}
\description{Data from a stratified random sample of size 300 from 
the U.S. 1992 Census of Agriculture.}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
437.}
