\name{anthrop}
\alias{anthrop}
\title{Length of Left Middle Finger and Height for 3000 Criminals...}
\description{Length of Left Middle Finger and Height for 3000 Criminals}
\details{Length of left middle finger and height for 3000 criminals}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
438.}
