\name{anthsrs}
\alias{anthsrs}
\title{Length of Left Middle Finger and Height for an SRS of Size 200...}
\description{Length of Left Middle Finger and Height for an SRS of Size 200}
\details{Length of left middle finger and height for an SRS of 200 
criminals from the anthrop dataset}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
438.}
