\name{anthuneq}
\alias{anthuneq}
\title{Length of Left Middle Finger and Height for an Unequal-Probability Sample of Size 200...}
\description{Length of Left Middle Finger and Height for an Unequal-Probability Sample of Size 200}
\details{Length of left middle finger and height for an unequal-probability sample 
of criminals of size 200 from the anthrop dataset. The probability of selection,
psi[i], was proportional to 24 for y < 65, 12 for y = 65, 2 for y = 66 or 67, 
and 1 for y > 67.}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
438.}
