\name{books}
\alias{books}
\title{Data from Home Owner's Survey on Total Number of Books...}
\description{Data from Home Owner's Survey on Total Number of Books}
\details{Data from home owner's survey on total number of books}
\docType{data}
\note{Used in Exercise 6 of Chapter 5.}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
439.}
