\name{brains}
\alias{brains}
\title{Relationship between brain and body weight}
\description{Average brain weights (g) and body weights (kg) for 62 species}
\format{
  Data frame with 62 rows and three variables:
\describe{
\item{\code{species}}{character}
\item{\code{body.wt}}{numeric}
\item{\code{brain.wt}}{numeric}
}
}
\source{Hand et al (1993) A Handbook of Small Data Sets. Chapman & Hall
}
\examples{
str(brains)
}
\keyword{datasets}
