\name{certify}
\alias{certify}
\title{Data from the 1994 Survey of ASA Membership on Certification...}
\description{Data from the 1994 Survey of ASA Membership on Certification}
\docType{data}
\note{The full dataset is on Statlib}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
439. 
\url{http://lib.stat.cmu.edu/asacert/certsurvey}}
