\name{chests}
\alias{chests}
\title{Chest sizes of 5738 Scottish soldiers}
\description{
Data frame with two variables:
\describe{
\item{\code{chest}}{chest measurement, numeric}
\item{\code{count}}{frequency, numeric}
 } 
}
\source{
Edinburgh Medical and Surgical Journal (1817)
Statements of the sizes of men in different counties of Scotland,
taken from the local militia *13* 260--264.
}
\examples{
head(chests)
}
\keyword{datasets}
