\name{cholesterol}
\alias{cholesterol}
\title{Effect of cholesterol on blood pressure}
\description{
Age, blood pressure and cholesterol were measured on a sample of 63 women.
}
\format{
  Data frame with 63 rows and three variables:
\describe{
\item{\code{age}}{age in years}
\item{\code{bp}}{blood pressure}
\item{\code{chol}}{serum cholesterol}
}
}
\source{Snedecor Statistical Methods. 5th edition 1956}

\examples{
head(cholesterol)
}
\keyword{datasets}
