\name{coots}
\alias{coots}
\title{Egg Size from Coots...}
\description{Egg Size from Coots}
\details{Selected information on egg size from coots, from a study
by Arnold (1991). Data courtesy of Todd Arnold.}
\docType{data}
\note{Not all observations are used for this data set, so
results may not agree with those in Arnold (1991)}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
440.}
