\name{darwin}
\alias{darwin}
\title{Darwin's maize plants}
\description{
Heights of 15 pairs of maize plants, one of each pair produced by 
crossing, the other by selfing. Unit = 1/8 inch.
}
\format{
Data frame with 15 rows and two variables:
\describe{
\item{\code{cross}}{height of the crossbred plant}
\item{\code{self}}{height of the selfed plant}
 }
}
\note{Observations are paired. Those in each row of the data frame represent
plants grown under similar conditions (e.g. in the same pot).
}
\source{
Fisher (1935) Design of Experiments. (Ch III)
}
\examples{
head(darwin)
}
\keyword{datasets}
