\name{divorce}
\alias{divorce}
\title{Data from a Sample of Divorce Records...}
\description{Data from a Sample of Divorce Records}
\details{Data from a sample of divorce records for states in the
Divorce Registration Area (National Center for Health Statistics 1987)}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
440.}
