\name{farms}
\alias{farms}
\title{Blood serum Cu levels in sheep}
\description{
Cu levels in 42 sheep sampled from 6 farms (seven sheep from each farm).
Three of the farms were on one soil type, three on another. 
}
\format{
A data frame with three variables:
\describe{
\item{\code{soil}}{factor}
\item{\code{farm}}{factor}
\item{\code{Cu}}{numeric}
}
}
\examples{
head(farms)
}
\keyword{datasets}
