\name{fingers}
\alias{fingers}
\title{Finger ridges}
\description{
The number of finger ridges in 12 pairs of identical twins
}
\format{Data frame with 24 observations, and two variables:
\describe{
 \item{\code{pair}}{(factor with 12 levels)}
 \item{\code{ridges}}{(numeric)}
  }
}
\source{
Snedecor Statistical Methods. 5th edition 1956
}
\examples{
head(fingers)
}
\keyword{datasets}
