\name{forbes}
\alias{forbes}
\title{Air pressure and boiling point of water}
\description{
Boiling point of water measured at 17 different barometric pressures
}
\format{
  Data frame with two variables:
\describe{
\item{\code{bp}}{boiling point (degrees F)}
\item{\code{pres}}{barometric pressure (inches of mercury)}
  }
}
\source{
The MASS package
}
\examples{
head(forbes)
}
\keyword{datasets}
