\name{galton}
\alias{galton}
\title{Heights of offspring and parents}
\description{
Adult height of offspring and the average height of the two parents in 928
families. All measurements rounded to nearest inch.
}
\format{
Data frame with three variables:
\describe{
\item{\code{parent}}{Average height of two parents}
\item{\code{child}}{Height of offspring}
\item{\code{frequency}}{Number of cases}
 }
}
\examples{
head(galton)
}
\keyword{datasets}
