\name{golfsrs}
\alias{golfsrs}
\title{Simple Random Sample of Golf Courses...}
\description{Simple Random Sample of Golf Courses}
\details{Simple Random Sample (SRS) of 120 golf courses taken from 
the population of the Website www.golfcourse.com}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
TODO.}
