\name{hills}
\alias{hills}
\title{Hill races}
\description{
Winning times for 35 hill races of various lengths and 'hilliness'.
}
\format{Data frame with 35 rows and 4 columns:
\describe{
\item{\code{race}}{factor}
\item{\code{time}}{numeric}
\item{\code{distance}}{numeric}
\item{\code{climb}}{numeric}
 }
}
\source{Hand et al (1993) A Handbook of Small Data Sets.
Chapman & Hall
}
\examples{
head(hills)
}
\keyword{datasets}
