\name{htstrat}
\alias{htstrat}
\title{Height and gender for a stratified random sample from htpop...}
\description{Height and gender for a stratified random sample from htpop}
\details{Height and gender for a stratified random sample of 160 women
and 40 men taken from the htpop population}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 230--234 and
442.}
