\name{irrigation}
\alias{irrigation}
\title{An irrigation experiment}
\description{
Crop yields for 24 plots of an irrigation experiment with 8 treatments 
}
\format{Data frame with 24 rows and two columns:
\describe{
\item{\code{treatment}}{factor}
\item{\code{yield}}{numeric}
 }
}
\examples{
head(irrigation)
}
\keyword{datasets}
