\name{measles}
\alias{measles}
\title{Survey of Parents of Children Non-Immunized against Measles...}
\description{Survey of Parents of Children Non-Immunized against Measles}
\details{Roberts et al. (1995) report on the results of a survey of parents whose
children had not been immunized against measles during a recent campaign
to immunize all children in the first five years of secondary school.}
\docType{data}
\note{The original data were unavailable; univariate and multivariate summary
statistics from these artificial data, however, are consistent with those
in the paper.}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
442.}
