\name{mice}
\alias{mice}
\title{
Weight distribution among 50 male mice
}
\description{
Weights at 60 days af 50 male mice. What is peculiar about the distribution?
}
\format{
A vector of length 50
}
\source{Falconer and Mackay Quantitative Genetics 4th ed, problem 6.2}
\examples{
summary(mice)
}
\keyword{datasets}
