\name{ncvs}
\alias{ncvs}
\title{Victimization Incidents in the July-December 1989 NCVS...}
\description{Victimization Incidents in the July-December 1989 NCVS}
\details{Selected variables for victimization incidents in the 
July-December 1989 NCVS. Note that some variables were 
recoded from the original data file.}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
443.}
