\name{nybight}
\alias{nybight}
\title{Data Collected in the New York Bight...}
\description{Data Collected in the New York Bight}
\details{Data collected in the New York Bight for June 1974 and 
June 1975 (Wilk et al. 1977)}
\docType{data}
\note{Two of the original strata were combined because of insufficient 
sample sizes.}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
443.}
