\name{otters}
\alias{otters}
\title{Otters Data...}
\description{Otters Data}
\details{Data on number of holts (dens) in Shetland, United Kingdom used
in Kruuk et al. (1989). (Data courtesy of Hans Kruuk).}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
443.}
