\name{ozone}
\alias{ozone}
\title{Ozone Readings from Eskdalemuir, for 1994 and 1995...}
\description{Ozone Readings from Eskdalemuir, for 1994 and 1995}
\details{Hourly ozone readings in parts per billion (ppb) from Eskdalemuir,
Scotland, for 1994 and 1995}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
443.}
