\name{rats}
\alias{rats}
\title{A food experiment on rats}
\description{
Twenty rats were used in an expt to compare two diets. Growth rate and
food intake were measured for each rat.
}
\format{Data frame with 20 rows and three variables:
\describe{
 \item{\code{diet}}{factor}
 \item{\code{x}}{food intake, numeric}
\item{\code{y}}{growth rate, numeric}
 }
}
\examples{
head(rats)
}
\keyword{datasets}
