\name{regressions}
\alias{regressions}
\title{Anscombe's regression example}
\description{
Three regressions with same intercept and slope but very
different residuals.
}
\format{
  Data frame with 11 rows and 4 four variables:
\describe{
\item{\code{x}}{(numeric)}
\item{\code{y1}}{(numeric)}
\item{\code{y2}}{(numeric)}
\item{\code{y3}}{(numeric)} 
  }
}
\source{Anscombe, FJ (1973) Graphs in statistical analysis. The
American Statistician *27*, 17--21.}
\examples{
head(regressions)
}
\keyword{datasets}
