\name{seals}
\alias{seals}
\title{Breathing Holes of Seals...}
\description{Breathing Holes of Seals}
\details{Data on number of breathing holes found in sampled areas of Svalbard 
fjords, reconstructed from summary statistics given in Lydersen and
Ryg (1991)}
\docType{data}
\note{The data are used in Chapter 4, Exercise 11.}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
444.}
