\name{selectrs}
\alias{selectrs}
\title{Steps used in Selecting an SRS...}
\description{Steps used in Selecting an SRS}
\details{Steps used in selecting the simple random sample (SRS) in Example 2.4
of Lohr(1999).}
\docType{data}
\note{the set of indices in column \code{e} was used to select
observations from \code{agpop} into dataset \code{agsrs}.}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 31--34 
and 444.}
