\name{sheep}
\alias{sheep}
\title{Serum Cu measured on three breeds of sheep}
\description{
An experiment to investigate differences in serum Cu among three breeds of
sheep.
}
\format{
Data frame with 15 rows and two variables:
\describe{
\item{\code{breed}}{factor with three levels}
\item{\code{cu}}{numeric}
}
}
\examples{
str(sheep)
}
\keyword{datasets}
