\name{statepop}
\alias{statepop}
\title{Unequal-Probability Sample of Counties in the US...}
\description{Unequal-Probability Sample of Counties in the US}
\details{counties selected with probability proportional to 1992 population}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 190  -- 192 and
444.}
