\name{statepps}
\alias{statepps}
\title{Information on States...}
\description{Information on States}
\details{Number of counties, land area, and population for the 50 states plus
the District of Columbia}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
445.}
