\name{syc}
\alias{syc}
\title{Survey of Youth in Custody, 1987...}
\description{Survey of Youth in Custody, 1987}
\details{The 1987 Survey of Youth in Custody sampled juveniles and young adults in 
long-term, state-operated juvenile institutions. Residents of facilities
at the end of 1987 were interviewed about family background, previous
criminal history, and drug and alcohol use. Selected variables from the
survey are contained in the syc data frame.}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. 235--239 and
445.}
