\name{teachers}
\alias{teachers}
\title{Elementary School Teacher Workload Data...}
\description{Elementary School Teacher Workload Data}
\details{Selected variables from a study on elementary school teacher workload
in Maricopa County, Arizona.}
\docType{data}
\note{The study is described in Exercise 16 of Chapter 15. The psu sizes
are given in \code{\link{teachmi}}. The large stratum had 245 schools; the 
small/medium stratum had 66 schools.}
\references{Gnap, R. (1995). Teacher load in Arizona elementary school 
districts in Maricopa County. Ph.D. diss., Arizona State University 

Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
446.}
