\name{teachmi}
\alias{teachmi}
\title{Cluster Sizes for Elementary School Teacher Workload Data...}
\description{Cluster Sizes for Elementary School Teacher Workload Data}
\details{Cluster sizes for the study on elementary school teacher workload
in Maricopa County, Arizona.}
\docType{data}
\note{The study is described in Exercise 16 of Chapter 15. The 
actual date are given in \code{\link{teachers}}.}
\references{Gnap, R. (1995). Teacher load in Arizona elementary school 
districts in Maricopa County. Ph.D. diss., Arizona State University 

Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
446.}
