\name{teachnr}
\alias{teachnr}
\title{Follow-Up Study of Nonrespondents from Gnap (1995)...}
\description{Follow-Up Study of Nonrespondents from Gnap (1995)}
\details{Follow-up study of nonrespondents from the Gnap (1995) study on the
workload of elementary school teachers in Maricopa County, Arizona.}
\docType{data}
\note{The study is described in Exercise 16 of Chapter 15. The 
actual date are given in \code{\link{teachers}}. Cluster size data for
the original study are given in \code{\link{teachmi}}.}
\references{Gnap, R. (1995). Teacher load in Arizona elementary school 
districts in Maricopa County. Ph.D. diss., Arizona State University 

Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
446.}
