\name{winter}
\alias{winter}
\title{ASU Winter Closure Survey...}
\description{ASU Winter Closure Survey}
\details{Selected variables from the Arizona State University Winter Closure
Survey, taken in January 1995. This survey was taken to investigate
the attitudes and opinions of university employees toward the closing
of the university between December 25 and January 1.}
\docType{data}
\references{Lohr (1999). Sampling: Design and Analysis, Duxbury, p. TODO and
447--448.}
